+++ src/cpu/options.cpp
@@ -4,11 +4,11 @@ namespace trimesh {
 
 using namespace pybind11::literals;
 
-float __global_tolerance = 1e-5;
+double __global_tolerance = 1e-5;
 
-void set_tolerance(float tolerance) { __global_tolerance = tolerance; }
+void set_tolerance(double tolerance) { __global_tolerance = tolerance; }
 
-float get_tolerance() { return __global_tolerance; }
+double get_tolerance() { return __global_tolerance; }
 
 void add_options_modules(py::module &m) {
     m.def("set_tolerance", &set_tolerance,
